gitogram: Handle already clonned chat
If createChatInfo() failed to commit, then remove chat info file and
return an error about it. If it failed when push to repo, try to reset
last commit and return an error about it. User at this step must fix the
error that caused AddChat() to fail.

If chat was already clonned but during add failed on some step
(may be in createChatInfo()), so it doesn't appear in Chats array,
proceed with it's adding and again return an error if something failed.

Issue: #30